requirements.txt
discord.py
beautifulsoup4
requests
python-dotenv
playwright
